Split out cargo module
